https://www.appbrewery.co/courses/851555/lectures/15448496
Now that you have learnt about Column and Row Widgets and how to use them to layout your app, it's time to test your understanding. Head over to the link below and download a layout that I would like you to try and replicate with code. I recommend using the existing Mi Card app to create this layout and once you've achieved the desired look and you're satisfied with the solution then you can delete the challenge code from the project.